单点登陆

+++ dd-manager/dd-manager-interface/src/main/java/com/ymd/ddshop/service/ItemService.java
@@ -18,5 +18,5 @@ public interface ItemService {
 
     int updateBatch(List<Long> ids);
 
-    int saveItem(TbItem tbItem, String content,String paramData);
+    Long saveItem(TbItem tbItem, String content,String paramData);
 }
